requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.0
requests>=2.31.0
google-generativeai>=0.3.0
django-cors-headers>=4.3.0
google-genai>=1.50.1
python-dotenv>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
